sparse: use an 'int' for 'id' field in node data
Making this a `double` seems to have been a mistake. Only `int` values are ever
stored into this field and it is only ever compared against `int` values.

+++ lib/sparse/QuadTree.c
@@ -21,7 +21,7 @@ extern double distance_cropped(double *x, int dim, int i, int j);
 struct node_data_struct {
   double node_weight;
   double *coord;
-  double id;
+  int id;
   void *data;
 };
